Fix layouts require cycles
