+++ index.js
@@ -2,7 +2,7 @@ import { Navigation } from 'react-native-navigation';
 import { YellowBox, UIManager } from 'react-native';
 
 import { HOC } from './App';
-import { setInitialLayout } from './src/navigation/layout';
+import { setInitialLayout } from './src/navigation/layouts/initialLayout';
 import { register } from './src/navigation/register';
 import { patchCustomConfig } from './src/features/debug/patch-custom-config';
