+++ src/navigation/layouts/chatLayout.ts
@@ -1,0 +1,9 @@
+import { CHAT_SCREEN } from '../screens/chat';
+
+export const getChatLayout = () => ({
+  root: {
+    stack: {
+      children: [CHAT_SCREEN],
+    },
+  },
+});
