+++ src/navigation/layouts/debugLayout.ts
@@ -1,0 +1,9 @@
+import { DEBUG_SCREEN } from 'src/navigation/screens/debug';
+
+export const getDebugLayout = () => ({
+  root: {
+    stack: {
+      children: [DEBUG_SCREEN],
+    },
+  },
+});
