+++ src/navigation/layouts/marketingLayout.ts
@@ -1,0 +1,9 @@
+import { MARKETING_SCREEN } from '../screens/marketing';
+
+export const getMarketingLayout = () => ({
+  root: {
+    stack: {
+      children: [MARKETING_SCREEN],
+    },
+  },
+});
