+++ src/navigation/layouts/newOfferLayout.ts
@@ -1,0 +1,9 @@
+import { NEW_OFFER_SCREEN } from 'src/navigation/screens/new-offer';
+
+export const getNewOfferLayout = () => ({
+  root: {
+    stack: {
+      children: [NEW_OFFER_SCREEN],
+    },
+  },
+});
