+++ src/navigation/layouts/utils.ts
@@ -1,0 +1,4 @@
+export const shouldShowDashboard = (insuranceStatus: string) =>
+  ['ACTIVE', 'INACTIVE_WITH_START_DATE', 'INACTIVE'].indexOf(
+    insuranceStatus,
+  ) !== -1;
